requirements.txt
numpy
pandas
matplotlib
pillow
tqdm
lmdb
albumentations